requirements.txt
Bootstrap-Flask==2.2.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
openai-whisper==20230314
python-dotenv==1.0.0
torch==2.0.0
yt-dlp==2023.3.4
